requirements.txt
Django==1.11.3
djangorestframework==3.6.3
Pillow==4.2.0
drfdocs==0.0.11
coreapi==2.3.1